// typescript/looker/sdk/models.ts
export interface IAccessToken {
  /**
   * Access Token used for API calls (read-only)
   */
  access_token?: string
  /**
   * Type of Token (read-only)
   */
  token_type?: string
  /**
   * Number of seconds before the token expires (read-only)
   */
  expires_in?: number
}
// ...
export interface IError {
  /**
   * Error details (read-only)
   */
  message: string
  /**
   * Documentation link (read-only)
   */
  documentation_url: Url
}